\name{SASxport-package}
\alias{SASxport-package}
\alias{SASxport}
\docType{package}
\title{
Read and write SAS export files
}
\description{
This package provides functions to read, list contents of,
and write SAS export files.
}
\details{

This package was created by Random
Technologies LLC <http://random-technologies-llc.com> with funding by
Metrum Institute <http://metruminstitute.org>.
  
Index:
\preformatted{
lookup.xport            Lookup Information on a SAS XPORT Format
                        Library
read.xport              Import SAS XPORT files
toSAS.default           Convert R data object for storage in SAS xport
                        file
units                   Set or Retreive the label, format, iformat, or
                        units Attribute of a Vector
write.xport             Write data to a SAS XPORT file
}
}
\author{
Gregory R. Warnes <greg@random-technologies-llc.com>, includes
code from Frank E. Harrell, Jr.'s Hmisc package.

Maintainer: Gregory R. Warnes <greg@random-technologies-llc.com>
}
\section{support}{
Technical support contracts and other services for for R, this package,
and other packages are available from Random Technologies LLC
<http://random-technologies-llc.com>.
}
\keyword{ package }
